+++ lib/src/disposable.dart
@@ -1,14 +1,25 @@
 /// Interface for objects you can dispose
 abstract class Disposable {
-  factory Disposable(void Function() dispose) => _DisposableImpl(dispose); 
+  static Disposable create<T> (T disposable, Function(T) dispose) =>
+    _DisposableImpl(disposable, dispose);
 
   void dispose();
 }
 
-class _DisposableImpl implements Disposable {
-  void Function() _dispose;
+class _DisposableImpl<T> implements Disposable {
+  T disposable;
+  void Function(T) _dispose;
 
-  _DisposableImpl(this._dispose) : assert(_dispose != null);
+  _DisposableImpl(this.disposable, this._dispose) :
+    assert(_dispose != null);
 
-  @override void dispose() => _dispose();
+  @override
+  bool operator ==(other) =>
+    (other is _DisposableImpl && other.disposable == disposable) ||
+      other == disposable;
+
+  @override
+  int get hashCode => disposable.hashCode;
+
+  @override void dispose() => _dispose(disposable);
 }