+++ lib/src/disposable_extensions.dart
@@ -1,23 +1,28 @@
 import 'dart:async';
 
-import 'package:disposable/src/disposable.dart';
-import 'package:disposable/src/disposable_collector.dart';
+import 'package:disposable_utils/src/disposable.dart';
+import 'package:disposable_utils/src/disposable_collector.dart';
 
 extension StreamControllerExtensions<T> on StreamController<T> {
-  Disposable toDisposable() => Disposable(() => this.close());
+  Disposable asDisposable() => Disposable.create(this, (s) => s.close());
 
   void addTo(DisposableCollector disposableCollection) =>
-    this.toDisposable().addTo(disposableCollection);
+    this.asDisposable().addTo(disposableCollection);
 }
 
 extension StreamSubscriptionExtensions<T> on StreamSubscription<T> {
-  Disposable toDisposable() => Disposable(() => this.cancel());
+  Disposable asDisposable() => Disposable.create(this, (s) => s.cancel());
 
   void addTo(DisposableCollector disposableCollection) =>
-    this.toDisposable().addTo(disposableCollection);
+    this.asDisposable().addTo(disposableCollection);
 }
 
 extension DisposableExtensions on Disposable {
   void addTo(DisposableCollector disposableCollector) =>
     disposableCollector.add(this);
+}
+
+extension ObjectExtensions<T> on T {
+  Disposable toDisposable (void Function(T) dispose) =>
+    Disposable.create(this, dispose);
 }