+++ test/disposable_test.dart
@@ -1,12 +1,11 @@
-import 'package:disposable/disposable.dart';
+import 'package:disposable_utils/disposable.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
 void main() {
   test('Disposable disposes the value', () {
     final a = BMock();
-
-    final d = new Disposable(() => a.close());
+    final d = a.toDisposable((a) => a.close());
     d.dispose();
 
     verify(a.close());