+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.2.0] 
+Added equality comparison and hashcode function to disposable.
+Added toDisposable(void Function(T)) to convert from any object
+toDisposable of streams and stream controllers removed to asDisposable() 
+Changed Disposable factory function to Disposable.create()
+
 ## [1.0.0]
 
 First release. Added Disposable, DisposableCollector and extension methods